+++ src/main/java/cart/application/OrderService.java
@@ -5,15 +5,15 @@ import cart.domain.MemberCoupon;
 import cart.domain.Order;
 import cart.domain.Product;
 import cart.domain.Products;
+import cart.domain.repository.MemberCouponRepository;
+import cart.domain.repository.OrderRepository;
+import cart.domain.repository.ProductRepository;
 import cart.domain.vo.Amount;
-import cart.dto.CartItemRequest;
-import cart.dto.OrderProductResponse;
-import cart.dto.OrderRequest;
-import cart.dto.OrderResponse;
+import cart.dto.request.CartItemRequest;
+import cart.dto.request.OrderRequest;
+import cart.dto.response.OrderProductResponse;
+import cart.dto.response.OrderResponse;
 import cart.exception.BusinessException;
-import cart.repository.MemberCouponRepository;
-import cart.repository.OrderRepository;
-import cart.repository.ProductRepository;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.stream.Collectors;
