+++ src/main/java/cart/domain/repository/CartItemRepository.java
@@ -1,4 +1,4 @@
-package cart.repository;
+package cart.domain.repository;
 
 import cart.domain.CartItem;
 import java.util.List;
