+++ src/main/java/cart/domain/repository/MemberCouponRepository.java
@@ -1,4 +1,4 @@
-package cart.repository;
+package cart.domain.repository;
 
 import cart.domain.MemberCoupon;
