+++ src/main/java/cart/dto/request/RemoveCartItemsRequest.java
@@ -1,4 +1,4 @@
-package cart.dto;
+package cart.dto.request;
 
 import java.util.List;
